pyproject.toml
[project]
name = "ml-platform"
version = "0.1.0"
description = "Shared ML platform: ClearML/Hydra contracts and workflow utilities"
requires-python = ">=3.10"
dependencies = [
  "clearml>=1.16.0",
  "hydra-core>=1.3.0",
  "omegaconf>=2.3.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "joblib>=1.3.0",
  "matplotlib>=3.7.0",
  "rich>=13.0.0",
  "pyyaml>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
